refactor: plugin path correction
