+++ packages/nocodb/src/plugins/linode/index.ts
@@ -5,7 +5,7 @@ import type { XcPluginConfig } from '~/types/nc-plugin';
 const config: XcPluginConfig = {
   builder: LinodeObjectStoragePlugin,
   id: 'linode',
-  fallbackTitle: 'Linode Object Storage',
+  recoveryTitle: 'Linode Object Storage',
   title: 'Linode',
   version: '0.0.4',
   logo: 'plugins/linode.svg',
