+++ packages/nocodb/src/plugins/minio/Minio.ts
