+++ packages/nocodb/src/plugins/r2/index.ts
@@ -6,7 +6,7 @@ const config: XcPluginConfig = {
   builder: R2Plugin,
   id: 'cloudflare-r2',
   title: 'Cloudflare R2',
-  fallbackTitle: 'Cloudflare R2 Storage',
+  recoveryTitle: 'Cloudflare R2 Storage',
   version: '0.0.3',
   logo: 'plugins/r2.png',
   description:
