+++ packages/nocodb/src/plugins/scaleway/index.ts
@@ -6,7 +6,7 @@ const config: XcPluginConfig = {
   builder: ScalewayObjectStoragePlugin,
   id: 'scaleway',
   title: 'Scaleway',
-  fallbackTitle: 'Scaleway Object Storage',
+  recoveryTitle: 'Scaleway Object Storage',
   version: '0.0.4',
   logo: 'plugins/scaleway.png',
   tags: 'Storage',
